# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "乐动数据库服务API文档",
        "title": "Ledong DB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:31168",
    "basePath": "/api",
    "paths": {
...
        "/user/": {
            "get": {
                "description": "根据会员编号查询单个会员，或查询所有会员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询会员列表或单个会员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会员编号",
                        "name": "number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PrepaidCard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/charged": {
            "post": {
                "description": "为用户充值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "充值",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "会员编号",
                        "name": "number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float32",
                        "description": "充值金额",
                        "name": "charged",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "充值时间",
                        "name": "time",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "format": "float32",
                        "description": "充值次数",
                        "name": "times",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "format": "float32",
                        "description": "年卡次数",
                        "name": "annualTimes",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "年卡过期时间",
                        "name": "annualExpireTime",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "等值",
                        "name": "worth",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "场地",
                        "name": "court",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "教练",
                        "name": "coach",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "备注",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Charge"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/charged/coach/{number}": {
            "get": {
                "description": "根据教练编号和时间范围查询充值记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询教练的充值记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "教练编号",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "endTime",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/charged/retreat/{id}": {
            "post": {
                "description": "根据充值记录ID进行退费",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "退费",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "充值记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/charged/total": {
            "get": {
                "description": "查询所有充值记录，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询所有充值记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ChargePage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/charged/{number}": {
            "get": {
                "description": "根据会员编号查询充值记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询某个用户的充值记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会员编号",
                        "name": "number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.ChargePage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/coach": {
            "get": {
                "description": "查询所有活跃的教练",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询教练列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Coach"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/course/notify": {
            "post": {
                "description": "发送课程通知短信",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "课程通知",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "课程ID，可选，不传则通知所有未通知的课程",
                        "name": "id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/court": {
            "get": {
                "description": "查询所有场地",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询场地列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Court"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/member/{number}": {
            "post": {
                "description": "设置会员的青少年和成人数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "设置会员的青少年和成人数量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会员编号",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "青少年数量",
                        "name": "yonth",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "成人数量",
                        "name": "adult",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PrepaidCard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "创建新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "注册用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "安全验证头",
                        "name": "secure",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "姓名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "会员编号",
                        "name": "number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "场地",
                        "name": "court",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，0表示成功",
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.PrepaidCard": {
            "type": "object",
            "properties": {
                "adults": {
                    "type": "integer"
                },
                "annual_count": {
                    "type": "number"
                },
                "annual_expire_time": {
                    "type": "string"
                },
                "charges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Charge"
                    }
                },
                "courses": {
                    "description": "关联关系",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Course"
                    }
                },
                "court": {
                    "type": "string"
                },
                "equivalent_balance": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "rest_charge": {
                    "type": "number"
                },
                "spends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Spend"
                    }
                },
                "times_count": {
                    "type": "number"
                },
                "times_expire_time": {
                    "type": "string"
                },
                "younths": {
                    "type": "integer"
                }
            }
        },
...
    }
}